Struct usrnet::core::arp_cache::ArpCache

pub struct ArpCache<T = SystemEnv> where
    T: Env,  { /* fields omitted */ }

Maintains an expiring set of IPv4 -> ethernet address mappings.

Methods

impl<T: Env> ArpCache<T>

pub fn new(expiration_in_secs: u64, time_env: T) -> ArpCache<T>

Creates an ARP cache where ethernet address mappings expire after expiration_in_secs seconds.

pub fn eth_addr_for_ip(
&mut self,
ipv4_addr: Ipv4Address
) -> Option<EthernetAddress>

Lookup the ethernet address for an IPv4 address.

pub fn set_eth_addr_for_ip(
&mut self,
ipv4_addr: Ipv4Address,
eth_addr: EthernetAddress
)

Create or update the ethernet address mapping for an IPv4 address.