Module usrnet::core::repr

Serialization and deserialization of network packets.

The repr module provides abstractions for serialization and deserializing packets and frames at different network layers to/from byte buffers.

Re-exports

pub use self::arp::hw_types as arp_hw_types;

pub use self::arp::proto_types as arp_proto_types;

pub use self::arp::Arp;

pub use self::arp::Op as ArpOp;

pub use self::ethernet::eth_types;

pub use self::ethernet::Address as EthernetAddress;

pub use self::ethernet::Frame as EthernetFrame;

pub use self::icmpv4::DestinationUnreachable as Icmpv4DestinationUnreachable;

pub use self::icmpv4::Message as Icmpv4Message;

pub use self::icmpv4::Packet as Icmpv4Packet;

pub use self::icmpv4::Repr as Icmpv4Repr;

pub use self::icmpv4::TimeExceeded as Icmpv4TimeExceeded;

pub use self::ipv4::flags as ipv4_flags;

pub use self::ipv4::protocols as ipv4_protocols;

pub use self::ipv4::Address as Ipv4Address;

pub use self::ipv4::AddressCidr as Ipv4AddressCidr;

pub use self::ipv4::Packet as Ipv4Packet;

pub use self::ipv4::Protocol as Ipv4Protocol;

pub use self::ipv4::Repr as Ipv4Repr;

pub use self::tcp::Control as TcpControl;

pub use self::tcp::Packet as TcpPacket;

pub use self::tcp::Repr as TcpRepr;

pub use self::tcp::SeqNum as TcpSeqNum;

pub use self::udp::Packet as UdpPacket;

pub use self::udp::Repr as UdpRepr;

Modules

arp

ethernet

icmpv4

ipv4

tcp

udp