Struct usrnet::core::repr::ipv4::Packet

pub struct Packet<T: AsRef<[u8]>> { /* fields omitted */ }

View of a byte buffer as an IPv4 packet.

Methods

impl<T: AsRef<[u8]>> Packet<T>

pub const MIN_HEADER_LEN: usize

MIN_HEADER_LEN: usize = 20

pub fn try_new(buffer: T) -> Result<Packet<T>>

Tries to create an IPv4 packet from a byte buffer.

NOTE: Use check_encoding() before operating on the packet if the provided buffer originates from a untrusted source such as a link.

pub fn buffer_len(payload_len: usize) -> usize

Returns the length of an IPv4 packet with the specified payload size.

pub fn check_encoding(&self) -> Result<()>

Checks if the packet has a valid encoding. This may include checksum, field consistency, etc. checks.

pub fn gen_header_checksum(&self) -> u16

Calculates the header checksum.

pub fn ip_version(&self) -> u8

pub fn header_len(&self) -> u8

pub fn dscp(&self) -> u8

pub fn ecn(&self) -> u8

pub fn packet_len(&self) -> u16

pub fn identification(&self) -> u16

pub fn flags(&self) -> u8

pub fn fragment_offset(&self) -> u16

pub fn ttl(&self) -> u8

pub fn protocol(&self) -> u8

pub fn header_checksum(&self) -> u16

pub fn src_addr(&self) -> Address

pub fn dst_addr(&self) -> Address

pub fn payload(&self) -> &[u8]

impl<T: AsRef<[u8]> + AsMut<[u8]>> Packet<T>

pub fn set_ip_version(&mut self, version: u8)

pub fn set_header_len(&mut self, header_len: u8)

pub fn set_dscp(&mut self, dscp: u8)

pub fn set_ecn(&mut self, ecn: u8)

pub fn set_packet_len(&mut self, packet_len: u16)

pub fn set_identification(&mut self, id: u16)

pub fn set_flags(&mut self, flags: u8)

pub fn set_fragment_offset(&mut self, frag_offset: u16)

pub fn set_ttl(&mut self, ttl: u8)

pub fn set_protocol(&mut self, protocol: u8)

pub fn set_header_checksum(&mut self, header_checksum: u16)

pub fn set_src_addr(&mut self, addr: Address)

pub fn set_dst_addr(&mut self, addr: Address)

pub fn payload_mut(&mut self) -> &mut [u8]