Struct usrnet::core::repr::tcp::Packet

pub struct Packet<T: AsRef<[u8]>> { /* fields omitted */ }

View of a byte buffer as a TCP packet.

Methods

impl<T: AsRef<[u8]>> Packet<T>

pub const MIN_HEADER_LEN: usize

MIN_HEADER_LEN: usize = 20

pub fn try_new(buffer: T) -> Result<Packet<T>>

Tries to create an TCP packet from a byte buffer.

NOTE: Use check_encoding() before operating on the packet if constructing a packet via a buffer originating from an untrusted source like a link.

pub fn buffer_len(payload_len: usize) -> usize

Returns the length of a TCP packet with no options and the specified payload size.

pub fn check_encoding(&self, ipv4_repr: &Ipv4Repr) -> Result<()>

Checks if the packet has a valid encoding. This may include checksum, field consistency, etc. checks.

pub fn gen_packet_checksum(&self, ipv4_repr: &Ipv4Repr) -> u16

Calculates the packet checksum.

pub fn src_port(&self) -> u16

pub fn dst_port(&self) -> u16

pub fn seq_num(&self) -> u32

pub fn ack_num(&self) -> u32

pub fn data_offset(&self) -> u8

pub fn ns(&self) -> bool

pub fn cwr(&self) -> bool

pub fn ece(&self) -> bool

pub fn urg(&self) -> bool

pub fn ack(&self) -> bool

pub fn psh(&self) -> bool

pub fn rst(&self) -> bool

pub fn syn(&self) -> bool

pub fn fin(&self) -> bool

pub fn window_size(&self) -> u16

pub fn checksum(&self) -> u16

pub fn urgent_pointer(&self) -> u16

pub fn options(&self) -> &[u8]

pub fn payload(&self) -> &[u8]

impl<T: AsRef<[u8]> + AsMut<[u8]>> Packet<T>

pub fn set_src_port(&mut self, port: u16)

pub fn set_dst_port(&mut self, port: u16)

pub fn set_seq_num(&mut self, seq_num: u32)

pub fn set_ack_num(&mut self, ack_num: u32)

pub fn set_data_offset(&mut self, data_offset: u8)

pub fn set_ns(&mut self, ns: bool)

pub fn set_cwr(&mut self, cwr: bool)

pub fn set_ece(&mut self, ece: bool)

pub fn set_urg(&mut self, urg: bool)

pub fn set_ack(&mut self, ack: bool)

pub fn set_psh(&mut self, psh: bool)

pub fn set_rst(&mut self, rst: bool)

pub fn set_syn(&mut self, syn: bool)

pub fn set_fin(&mut self, fin: bool)

pub fn set_window_size(&mut self, window_size: u16)

pub fn set_checksum(&mut self, checksum: u16)

pub fn set_urgent_pointer(&mut self, urgent_pointer: u16)

pub fn options_mut(&mut self) -> &mut [u8]

pub fn payload_mut(&mut self) -> &mut [u8]

pub fn fill_checksum(&mut self, ipv4_repr: &Ipv4Repr)