Function usrnet::core::service::arp::eth_addr_for_ip

pub fn eth_addr_for_ip(
    interface: &mut Interface, 
    ipv4_addr: Ipv4Address
) -> Result<EthernetAddress>

Tries to retrieve the Ethernet address for an IPv4 address.

The IP address may not have an Ethernet mapping yet, in which case an ARP request is dispatched and an error returned. The ARP response (if the IP address exists on the network) will be processed by recv_packet(...) and update the ARP cache.