Function usrnet::core::service::ipv4::send_packet_raw

pub fn send_packet_raw<F>(
    interface: &mut Interface, 
    dst_addr: Ipv4Address, 
    ipv4_packet_len: usize, 
    f: F
) -> Result<()> where
    F: FnOnce(&mut [u8]), 

Send a raw IPv4 packet via the interface.

The appropriate Ethernet destination address will be inferred by the network stack, but the callers is responsible for writing an entire well formatted IPv4 packets to the provided buffer, NOT just the payload!