Struct usrnet::core::service::Interface

pub struct Interface {
    pub dev: Box<Device>,
    pub arp_cache: ArpCache,
    pub ethernet_addr: EthernetAddress,
    pub ipv4_addr: Ipv4AddressCidr,
    pub default_gateway: Ipv4Address,
    pub loopback: SliceBuf<u8>,
}

An interface for sending and receiving network packets.

Fields
dev: Box<Device>

Device for sending and receiving raw Ethernet frames.

arp_cache: ArpCache

Cache for IPv4/Ethernet address translations.

ethernet_addr: EthernetAddress

Ethernet address for the interface.

ipv4_addr: Ipv4AddressCidr

IPv4 address for the interface.

default_gateway: Ipv4Address

Default gateway for IPv4 packets not on the interface subnet. This should be on the same subnet as ipv4_addr!

loopback: SliceBuf<u8>

A set of of Ethernet frames captured for sending back up the stack.