Module usrnet::core::socket::env

Structs

SocketEnv
An environment for creating sockets configured for a particular interface.

Constants

RAW_SOCKET_BUFFER_LEN
Default capacity of a raw socket in bytes.

RAW_SOCKET_PACKETS
Default capacity of a raw socket in packets.

TCP_CONNECTION_QUEUE_LEN
Default number of connections a TCP listener can buffer.

TCP_RETRANSMISSION_TIMEOUT
Default TCP retransmission timeout in seconds.

TCP_SEND_QUEUE_LEN
Length in bytes of the outbound data stream queue.

TCP_SYN_RECV_TIMEOUT
Default timeout for dropping SYN_RECV states in seconds.

UDP_SOCKET_BUFFER_LEN
Default capacity of a UDP socket in bytes.

UDP_SOCKET_PACKETS
Default capacity of a UDP socket in packets.