Struct usrnet::core::socket::raw::RawSocket

pub struct RawSocket<'a> { /* fields omitted */ }

Socket for sending and receiving raw ethernet or IP packets.

Methods

impl<'a> RawSocket<'a>

pub fn new(
raw_type: RawType,
send_buffer: Ring<'a, Slice<'a, u8>>,
recv_buffer: Ring<'a, Slice<'a, u8>>
) -> RawSocket<'a>

Creates a socket with the provided send and receive buffers.

pub fn send(&mut self, buffer_len: usize) -> Result<&mut [u8]>

Enqueues a packet with buffer_len bytes for sending.

Errors

An Error::Exhausted occurs if the send buffer is full.

pub fn recv(&mut self) -> Result<&[u8]>

Dequeues a received packet from the socket.

Errors

An Error::Exhausted occurs if the receive buffer is full.