Trait usrnet::core::socket::socket::Socket

pub trait Socket {
    fn send_forward<F, R>(&mut self, f: F) -> Result<R>
    where
        F: FnOnce(Packet) -> Result<R>;

    fn recv_forward(&mut self, packet: &Packet) -> Result<()>;
}

A generic interface for processing socket packets.

Required Methods

fn send_forward<F, R>(&mut self, f: F) -> Result<R> where
F: FnOnce(Packet) -> Result<R>,

Attempts to dequeue a packet enqueued for sending via a function f which can process the packet. If f returns an error, the packet should NOT be dequeued.

Errors

An Error::Exhausted error occurs if the socket has no pending packets or another error if f returns an error.

fn recv_forward(&mut self, packet: &Packet) -> Result<()>

Provides a packet which the socket MAY enqueue for dequeueing in the future. The packet may be ignored for reasons including an exhausted buffer or disintrest in the packet.

Errors

An Error::Exhausted error occurs if the socket buffer is full or Error::NoOp if the packet was ignored.