Struct usrnet::core::socket::TcpSocket

pub struct TcpSocket { /* fields omitted */ }

A TCP stream for reliable sequential transfers created by establishing new client side connections or accepting connections through a listener.

Methods

impl TcpSocket

pub fn new<F>(
binding: SocketAddrLease,
rng: Box<Rng>,
time_env: Box<TimeEnv>,
config: TcpConfig,
callback: F
) -> TcpSocket where
F: 'static + FnMut(&mut TcpSocket, TcpEvent),

Creates a new TCP socket.

pub fn listen(&mut self)

Begins listening for connections.

Panics

Causes a panic if the socket is not in a closed state.

pub fn connect(&mut self, remote_addr: SocketAddr)

Begins establishing a connection to an endpoint.

Panics

Causes a panic if the connection is not in a closed state.

pub fn send(&mut self, buffer: &[u8]) -> Result<usize>

Copies the provided buffer into the outbound data stream for transmission.

Returns the number of bytes copied, which may be less than the provided buffer is the stream's send buffer is at or near capacity. An error is returned if the socket is closed.

pub fn dequeue<F>(&mut self, ipv4_packet_len_max: usize, f: F) -> Result<()> where
F: FnOnce(&Ipv4Repr, &TcpRepr, &[u8]) -> Result<()>,

Dequeues a packet enqueued for sending via a function f.

The packet is only dequeued if f does not return an error.

pub fn enqueue(
&mut self,
ipv4_repr: &Ipv4Repr,
tcp_repr: &TcpRepr,
payload: &[u8]
) -> Result<()>

Enqueues a packet for receiving.

This includes not just packets for data transfer, but 3-way handshake establishment, closing, etc.

pub fn accepts(&self, ipv4_repr: &Ipv4Repr, tcp_repr: &TcpRepr) -> bool

Checks if the stream is interested in a TCP packet.

The packet may be ultimately rejected for reasons such as a full receive buffer, bad sequence number, etc.

pub fn dispatch(&mut self)

Triggers any ready events. See TcpEvent for more info.