Struct usrnet::core::socket::TcpStream

pub struct TcpStream { /* fields omitted */ }

A TCP stream for reliable sequential transfers created by establishing new client side connections or accepting connections through a listener.

Methods

impl TcpStream

pub fn new(context: TcpContext) -> TcpStream

Creates a new closed TCP stream.

pub fn send(&mut self, buffer: &[u8]) -> Result<usize>

Copies the provided buffer into the outbound data stream for transmission.

Returns the number of bytes copied, which may be less than the provided buffer is the stream's send buffer is at or near capacity. An error is returned if the socket is closed.

pub fn connect(&mut self, remote_addr: SocketAddr)

Begins establishing a connection to an endpoint.

Panics

Causes a panic if the connection is not in a closed state.

pub fn state(&self) -> TcpState

Returns the current TCP state of the socket.

pub fn is_establishing(&self) -> bool

Checks if the socket is establishing a connection.