Struct usrnet::core::socket::TcpTimer

pub struct TcpTimer { /* fields omitted */ }

A timer with exponential backoff.

Methods

impl TcpTimer

pub fn new(delay: Duration, time_env: Box<TimeEnv>) -> TcpTimer

Creates a TcpTimer.

The delay specified the time between retries. This delay is compounded with an exponential backoff factor.

pub fn triggered(&mut self) -> bool

Checks if the timer has triggered.

WARNING: This function increases the backoff factor and will trigger again once the timeout runs out. In other words, you need to manually cancel the timer when using this.

pub fn schedule_delayed(&mut self)

Cancels an existing timer, schedules a delayed trigger of the timer, and resets the backoff factor.

pub fn schedule_now(&mut self)

Schedules an immediate trigger of the timer and resets the backoff factor.

pub fn reset(&mut self)

Cancels an existing timer and resets the backoff factor.

pub fn now(&self) -> Instant

Returns the current time.