Enum usrnet::core::socket::tagged::TaggedSocket

pub enum TaggedSocket {
    Raw(RawSocket),
    Udp(UdpSocket),
    TcpListener(TcpListener),
    TcpStream(TcpStream),
}

One of many types of sockets.

Variants
Raw(RawSocket)Udp(UdpSocket)TcpListener(TcpListener)TcpStream(TcpStream)
Methods

impl TaggedSocket

pub fn as_raw_socket(&mut self) -> &mut RawSocket

Returns a reference to the underlying raw socket.

Panics

Panics if the underlying socket is not a raw socket.

pub fn as_tcp_listener(&mut self) -> &mut TcpListener

Returns a reference to the underlying TCP listener.

Panics

Panics if the underlying socket is not a TCP listener.

pub fn as_tcp_stream(&mut self) -> &mut TcpStream

Returns a reference to the underlying TCP stream.

Panics

Panics if the underlying socket is not a TCP stream.

pub fn as_udp_socket(&mut self) -> &mut UdpSocket

Returns a reference to the underlying UDP socket.

Panics

Panics if the underlying socket is not a UDP socket.