Trait usrnet::core::socket::tcp::state::Tcp

pub trait Tcp<T: TimeEnv>: Into<TcpState<T>> {
    fn send_forward<F, R>(self, _: F) -> (TcpState<T>, Result<R>)
    where
        F: FnOnce(Packet) -> Result<R>,
    { ... }

    fn recv_forward(self, _: &Packet) -> (TcpState<T>, Result<()>) { ... }
}

A generic interface for implementing TCP state behavior and transitions.

Provided Methods

fn send_forward<F, R>(self, _: F) -> (TcpState<T>, Result<R>) where
F: FnOnce(Packet) -> Result<R>,

Similar to Socket::send_forward(...). In addition to the result of f, the next state (or same but updated state) of the TCP FSM is returned.

fn recv_forward(self, _: &Packet) -> (TcpState<T>, Result<()>)

Similar to Socket::recv_forward(...). In addition to a packet processing result, the next state (or same but updated state) of the TCP FSM is returned.