Struct usrnet::core::socket::tcp::TcpListen

pub struct TcpListen {
    pub syn_queue: VecDeque<TcpSynRecv>,
    pub est_queue: VecDeque<TcpEstablished>,
    pub context: TcpContext,
}

The TCP LISTENING state.

Fields
syn_queue: VecDeque<TcpSynRecv> est_queue: VecDeque<TcpEstablished> context: TcpContext
Methods

impl TcpListen

pub fn accept(&mut self) -> Option<TcpEstablished>

Dequeues an established connection if one exists.

pub fn recv_enqueue_syn(
&mut self,
ipv4_repr: &Ipv4Repr,
tcp_repr: &TcpRepr,
payload: &[u8]
) -> bool

Forwards a packet to an ESTABLISHED state.

Returns a boolean indicating if the packet was acceptable by any sockets.

pub fn recv_enqueue_est(
&mut self,
ipv4_repr: &Ipv4Repr,
tcp_repr: &TcpRepr,
payload: &[u8]
) -> bool

Forwards a packet to a SYN_RECV state.

Returns a boolean indicating if the packet was acceptable by any sockets.

pub fn to_syn_recv(
&mut self,
connecting_to: SocketAddr,
remote_seq_num: u32
) -> TcpSynRecv

Transitions from LISTEN to SYN_RECV in order to establish a new connection.