Struct usrnet::core::socket::tcp::TcpSocket

pub struct TcpSocket<T: TimeEnv = SystemEnv> { /* fields omitted */ }

A TCP socket.

Methods

impl<T: TimeEnv> TcpSocket<T>

pub fn new(
binding: SocketAddrLease,
time_env: T,
interface_mtu: usize
) -> TcpSocket<T>

Creates a new TCP socket.

pub fn connect(&mut self, socket_addr: SocketAddr)

Initiates a connection to a TCP endpoint.

Panics

Causes a panic if the connection is not in the closed state!

pub fn is_closed(&self) -> bool

Checks if the socket is closed. The socket may be closed for reasons including an explicit close, timeout, reset, etc.

pub fn is_establishing(&self) -> bool

Checks if the socket is connecting to an endpoint.

pub fn is_connected(&self) -> bool

Checks if the socket has connected to an endpoint.