Trait usrnet::core::socket::tcp::Tcp

pub trait Tcp {
    fn send_dequeue<F, R>(&mut self, _f: F) -> (Option<TcpState>, Result<R>)
    where
        F: FnOnce(&Ipv4Repr, &TcpRepr, &[u8]) -> Result<R>,
    { ... }

    fn recv_enqueue(
        &mut self, 
        _ipv4_repr: &Ipv4Repr, 
        _tcp_repr: &TcpRepr, 
        _payload: &[u8]
    ) -> (Option<TcpState>, Result<()>) { ... }
}

A generic interface for implementing TCP state behavior and transitions.

Provided Methods

fn send_dequeue<F, R>(&mut self, _f: F) -> (Option<TcpState>, Result<R>) where
F: FnOnce(&Ipv4Repr, &TcpRepr, &[u8]) -> Result<R>,

Dequeues a packet enqueued for sending via function f.

The packet is only dequeued if f does not return an error. In addition, returns the next TCP state the socket should transition to.

fn recv_enqueue(
&mut self,
_ipv4_repr: &Ipv4Repr,
_tcp_repr: &TcpRepr,
_payload: &[u8]
) -> (Option<TcpState>, Result<()>)

Enqueues a packet for receiving.

In addition, returns the next TCP state the socket should transition to.