Struct usrnet::core::socket::udp::UdpSocket

pub struct UdpSocket { /* fields omitted */ }

A UDP socket.

Methods

impl UdpSocket

pub fn new(
binding: SocketAddrLease,
send_buffer: SliceBuf<u8, SocketAddr>,
recv_buffer: SliceBuf<u8, SocketAddr>
) -> UdpSocket

Creates a new UDP socket.

pub fn send(&mut self, buffer_len: usize, addr: SocketAddr) -> Result<&mut [u8]>

Enqueues a packet with a payload_len bytes payload for sending to the specified address.

pub fn recv(&mut self) -> Result<(&[u8], SocketAddr)>

Dequeues a received packet along with it's source address from the socket.

pub fn send_dequeue<F, R>(&mut self, f: F) -> Result<R> where
F: FnOnce(&Ipv4Repr, &UdpRepr, &[u8]) -> Result<R>,

Dequeues a packet enqueued for sending via function f.

The packet is only dequeued if f does not return an error.

pub fn recv_enqueue(
&mut self,
ipv4_repr: &Ipv4Repr,
udp_repr: &UdpRepr,
payload: &[u8]
) -> Result<()>

Enqueues a packet for receiving.

pub fn send_enqueued(&self) -> usize

Returns the number of packets enqueued for sending.

pub fn recv_enqueued(&self) -> usize

Returns the number of packets enqueued for receiving.