Struct usrnet::core::storage::ring::Ring

pub struct Ring<'a, T: 'a> { /* fields omitted */ }

Ring/bounded buffer of T's.

Methods

impl<'a, T> Ring<'a, T>

pub fn dequeue_with<'b, F, R>(&'b mut self, f: F) -> Result<R> where
F: FnOnce(&'b mut T) -> R,

Applies f on the head of the buffer or returns an error if the buffer is empty. Dequeue's the element f was applied on.

Returns

An error or the result of f.

pub fn dequeue_maybe<'b, F, R>(&'b mut self, f: F) -> Result<R> where
F: FnOnce(&'b mut T) -> Result<R>,

Similar to dequeue_with(...) except cancels the dequeue operation if f returns an error.

Returns

An error or the result of f.

pub fn enqueue_with<'b, F, R>(&'b mut self, f: F) -> Result<R> where
F: FnOnce(&'b mut T) -> R,

Applies f on the head of the buffer (so that f can mutate the T as desired) or returns an error if the buffer is full. Enqueue's the element f was applied on.

Returns

An error or the result of f.

pub fn enqueue_maybe<'b, F, R>(&'b mut self, f: F) -> Result<R> where
F: FnOnce(&'b mut T) -> Result<R>,

Similar to dequeue_with(...) except cancels the enqueue operation if f returns an error.

Returns

An error or the result of f.

pub fn len(&self) -> usize

Returns the current number of items in the ring.