Struct usrnet::core::storage::slice_buf::SliceBuf

pub struct SliceBuf<T, M> { /* fields omitted */ }

A FIFO ordered sequence of T slices from a shared buffer, each with an associated piece of metadata M.

Methods

impl<T: Default, M: Default> SliceBuf<T, M>

pub fn new(buf_capacity: usize, slice_capacity: usize) -> SliceBuf<T, M>

Creates a new slice buffer.

The resulting buffer can store at at most buf_capacity T's and slice_capacity T slices. However depending on the alignment of slices within the buffer, enqueueing a slice may fail if the available capacity of the buffer is split among the head and tail. To avoid this, allocate at least 2 x the minimum buffer capacity you need.

impl<T: Copy, M> SliceBuf<T, M>

pub fn copy_from_slice(&mut self, buf: &[T], meta: M) -> Result<()>

Copies and creates a new slice at the end of the buffer.

impl<T, M> SliceBuf<T, M>

pub fn dequeue(&mut self) -> Result<(&mut [T], &mut M)>

Returns a slice of T's and the associated metadata if one exists.

pub fn dequeue_with<'a, F, R>(&'a mut self, f: F) -> Result<R> where
F: FnOnce(&'a mut [T], &'a mut M) -> R,

Applies f on a slice of T's and the associated metadata or returns an error if the buffer is empty. Dequeue's the slice f was applied on.

pub fn dequeue_maybe<'a, F, R>(&'a mut self, f: F) -> Result<R> where
F: FnOnce(&'a mut [T], &'a mut M) -> Result<R>,

Similar to dequeue_with(...) except cancels the dequeue operation if f returns an error.

pub fn enqueue_with<'a, F, R>(
&'a mut self,
buffer_len: usize,
meta: M,
f: F
) -> Result<R> where
F: FnOnce(&'a mut [T]) -> R,

Applies f on a slice of buffer_len T's so that f can mutate the buffer as desired. The provided metadata M is associated with the buffer if enqueueing is succesful.

pub fn enqueue_maybe<'a, F, R>(
&'a mut self,
buffer_len: usize,
meta: M,
f: F
) -> Result<R> where
F: FnOnce(&'a mut [T]) -> Result<R>,

Similar to enqueue_with(...) except cancels the enqueue operation if f returns an error.

pub fn slices(&self) -> usize

Returns the number of enqueued slices.