Enum usrnet::Error

pub enum Error {
    MacResolution(Ipv4Address),
    BindingInUse(SocketAddr),
    Exhausted,
    Ignored,
    Device(Option<IOError>),
    Malformed,
    Checksum,
}

Variants
MacResolution(Ipv4Address)

Indicates an error where a MAC address could not be resolved for an IPV4 address.

BindingInUse(SocketAddr)

Indicates an error where a socket binding has already been assigned.

Exhausted

Indicates an error where a socket buffer is full or empty, depending on the operation being performed.

Ignored

Indicates an error where a an incoming packet was ignored.

Device(Option<IOError>)

Indicates an error with a device/interface. This includes situations such as writes to a busy device or attempting reads on a device with no Ethernet frames.

Malformed

Indicates an error where a packet or frame is malformed.

Checksum

Indicates an error where a checksum is invalid.