Module usrnet::examples::env

Structs

DEFAULT_ETH_ADDR
Default interface MAC address.

DEFAULT_IPV4_ADDR
Default interface IPv4 address.

DEFAULT_IPV4_ADDR_CIDR
Default interface IPv4 address with a subnet mask.

DEFAULT_IPV4_GATEWAY
Default interface IPv4 gateway.

NO_HOST_IPV4_ADDR
An IPv4 address not assigned to any hosts on the network.

Statics

RAW_SOCKET_BUFFER_LEN
Default number of packets a raw socket can buffer.

SOCKET_SET_HANDLES
Default number of handles a socket set has capacity for.

UDP_SOCKET_BUFFER_LEN
Default number of packets a UDP socket can buffer.

Functions

default_dev

default_interface
Creates a network interface.

ifr_addr
Get's the IPv4 address for an interface. See tap.sh for more info.

raw_socket
Creates a raw socket.

ring
Creates a ring of len elements using factory function f.

socket_set
Creates a socket set.

tcp_socket
Creates a TCP socket.

tick
Sends and receives packets from/to sockets and the interface.

udp_socket
Creates a UDP socket.