Struct usrnet::core::repr::ethernet::Frame

pub struct Frame<T: AsRef<[u8]>> { /* fields omitted */ }

View of a byte buffer as an ethernet frame.

Methods

impl<T: AsRef<[u8]>> Frame<T>

pub const HEADER_LEN: usize

HEADER_LEN: usize = 14

pub const MAX_FRAME_LEN: usize

MAX_FRAME_LEN: usize = 1518

pub fn try_new(buffer: T) -> Result<Frame<T>>

Tries to create an ethernet frame from a byte buffer.

pub fn buffer_len(payload_len: usize) -> usize

Returns the length of an ethernet frame with the specified payload size.

pub fn dst_addr(&self) -> Address

pub fn src_addr(&self) -> Address

pub fn payload_type(&self) -> u16

pub fn payload(&self) -> &[u8]

impl<T: AsRef<[u8]> + AsMut<[u8]>> Frame<T>

pub fn set_dst_addr(&mut self, addr: Address)

pub fn set_src_addr(&mut self, addr: Address)

pub fn set_payload_type(&mut self, payload_type: u16)

pub fn payload_mut(&mut self) -> &mut [u8]