Struct usrnet::core::socket::raw::RawSocket

pub struct RawSocket { /* fields omitted */ }

Socket for sending and receiving raw ethernet or IP packets.

Methods

impl RawSocket

pub fn new<F>(
raw_type: RawType,
capacity_bytes: usize,
capacity_packets: usize,
callback: F
) -> RawSocket where
F: 'static + FnMut(&mut RawSocket, RawEvent),

Creates a raw socket with the specified capacity.

See RawEvent for more info on when the callback is invoked.

pub fn send(&mut self, buffer_len: usize) -> Result<&mut [u8]>

Sends an ethernet frame or IP packet with buffer_len bytes.

pub fn recv(&mut self) -> Result<&[u8]>

Receives an ethernet frame or IP packet.

pub fn dequeue<F, R>(&mut self, f: F) -> Result<R> where
F: FnOnce(&[u8]) -> Result<R>,

Dequeues a packet enqueued for sending via a function f.

The packet is only dequeued if f does not return an error.

pub fn enqueue(&mut self, packet: &[u8]) -> Result<()>

Enqueues a packet for receiving.

pub fn raw_type(&self) -> RawType

Returns the type of raw packets this socket contains.

pub fn dispatch(&mut self)

Triggers any ready events. See RawEvent for more info.