Struct usrnet::core::socket::set::SocketSet

pub struct SocketSet { /* fields omitted */ }

A set of sockets with stable integral handles.

Methods

impl SocketSet

pub fn new(socket_capacity: usize) -> SocketSet

Creates a socket set supporting a maximum number of sockets.

pub fn add_socket(&mut self, socket: TaggedSocket) -> Option<usize>

Adds a socket and returns a stable handle.

pub fn socket(&mut self, socket_handle: usize) -> &mut TaggedSocket

Returns a reference to a socket with the specified handle. Causes a panic if the handle is not in use.

pub fn iter_mut(
&mut self
) -> impl Iterator<Item = &mut TaggedSocket>

Returns an iterator over all of the sockets in the set.

This function returns an iterator which always starts at the element after the last one seen in a previous use of the iterator. Regardless of where it starts the iterator will traverse all of the sockets. This round-robin behavior is important for avoiding starvation when sending packets from sockets.