Enum usrnet::core::socket::tagged::TaggedSocket

pub enum TaggedSocket {
    Raw(RawSocket),
    Udp(UdpSocket),
    Tcp(TcpSocket),
    // some variants omitted
}

One of many types of sockets.

Variants
Raw(RawSocket)Udp(UdpSocket)Tcp(TcpSocket)
Methods

impl TaggedSocket

pub fn as_raw_socket(&mut self) -> &mut RawSocket

Returns a reference to the underlying raw socket.

Panics

Panics if the underlying socket is not a raw socket.

pub fn as_tcp_socket(&mut self) -> &mut TcpSocket

Returns a reference to the underlying TCP socket.

Panics

Panics if the underlying socket is not a TCP socket.

pub fn as_udp_socket(&mut self) -> &mut UdpSocket

Returns a reference to the underlying UDP socket.

Panics

Panics if the underlying socket is not a UDP socket.