Enum usrnet::core::socket::tagged::TaggedSocket

pub enum TaggedSocket<'a> {
    Raw(RawSocket<'a>),
    Udp(UdpSocket<'a>),
    // some variants omitted
}

One of many types of sockets.

Variants
Raw(RawSocket<'a>)Udp(UdpSocket<'a>)
Methods

impl<'a> TaggedSocket<'a>

pub fn as_raw_socket(&mut self) -> &mut RawSocket<'a>

Returns a reference to the underlying raw socket.

Panics

Panics if the underlying socket is not a raw socket.

pub fn as_udp_socket(&mut self) -> &mut UdpSocket<'a>

Returns a reference to the underlying UDP socket.

Panics

Panics if the underlying socket is not a UDP socket.