Struct usrnet::core::storage::ring::Ring

pub struct Ring<T> { /* fields omitted */ }

Ring/bounded buffer of T's.

Methods

impl<T> Ring<T>

pub fn peek(&self) -> Option<&T>

Returns a reference to the head of the ring if one exists.

pub fn dequeue(&mut self) -> Result<&mut T>

Dequeues an item from the ring if one exists.

pub fn dequeue_with<'a, F, R>(&'a mut self, f: F) -> Result<R> where
F: FnOnce(&'a mut T) -> R,

Applies f on the head of the buffer or returns an error if the buffer is empty. Dequeue's the element f was applied on.

pub fn dequeue_maybe<'a, F, R>(&'a mut self, f: F) -> Result<R> where
F: FnOnce(&'a mut T) -> Result<R>,

Similar to dequeue_with(...) except cancels the dequeue operation if f returns an error.

pub fn enqueue(&mut self, value: T) -> Result<()>

Enqueues an item onto the ring if capacity permits.

pub fn enqueue_with<'a, F, R>(&'a mut self, f: F) -> Result<R> where
F: FnOnce(&'a mut T) -> R,

Applies f on the head of the buffer (so that f can mutate the T as desired) or returns an error if the buffer is full. Enqueue's the element f was applied on.

pub fn enqueue_maybe<'a, F, R>(&'a mut self, f: F) -> Result<R> where
F: FnOnce(&'a mut T) -> Result<R>,

Similar to enqueue_with(...) except cancels the enqueue operation if f returns an error.

pub fn clear(&mut self)

Clears the ring of all items.

pub fn len(&self) -> usize

The current number of items in the ring.

pub fn avail(&self) -> usize

The number of slots available for items in the ring.

impl<T: Copy> Ring<T>

pub fn copy_to_slice(&self, buffer: &mut [T], offset: usize) -> usize

Copies as many T's as possible at the specified offset from the head of the ring into the provided buffer.

Returns the number of T's successfully copied.

pub fn enqueue_slice(&mut self, buffer: &[T]) -> usize

Copies the provided buffer into the ring.

Returns the number of T's succesfully copied.