Module usrnet::examples::env

Structs

DEFAULT_ETH_ADDR
Default interface MAC address.

DEFAULT_IPV4_ADDR
Default interface IPv4 address.

DEFAULT_IPV4_ADDR_CIDR
Default interface IPv4 address with a subnet mask.

DEFAULT_IPV4_GATEWAY
Default interface IPv4 gateway.

NO_HOST_IPV4_ADDR
An IPv4 address not assigned to any hosts on the network.

Constants

LOOPBACK_BUFFER_LEN
Size of our loopback frame buffer.

LOOPBACK_MAX_FRAMES
Maximum number of frames to enqueue on the loopback buffer.

SOCKET_SET_HANDLES
Default capacity of a socket set.

Functions

default_dev

default_interface
Creates a network interface.

ifr_addr
Get's the IPv4 address for an interface. See tap.sh for more info.

socket_env
Creates a socket environment.

socket_set
Creates a socket set.

tick
Sends and receives packets from/to sockets and the interface.